+++ packages/starrating/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.3.4](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-starrating@7.3.3...@pluralsight/ps-design-system-starrating@7.3.4) (2020-11-24)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-starrating
+
+
+
+
+
 ## [7.3.3](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-starrating@7.3.2...@pluralsight/ps-design-system-starrating@7.3.3) (2020-11-18)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-starrating